+++ js/init.js
@@ -1,0 +1,38 @@
+const ARRAY_USERS_COUNT = 10;
+
+const PLACE_AND_TIME = {
+  apartments:
+  [
+    'palace',
+    'flat',
+    'house',
+    'bungalow',
+    'hotel',
+  ],
+  time:
+  [
+    '12:00',
+    '13:00',
+    '14:00',
+  ],
+};
+
+const FACILITIES = {
+  additions:
+  [
+    'wifi',
+    'dishwasher',
+    'parking',
+    'washer',
+    'elevator',
+    'conditioner'
+  ],
+  interior:
+  [
+    'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/duonguyen-8LrGtIxxa4w.jpg',
+    'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/brandon-hoogenboom-SNxQGWxZQi0.jpg',
+    'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/claire-rendall-b6kAwr1i0Iw.jpg'
+  ],
+};
+
+export {ARRAY_USERS_COUNT, PLACE_AND_TIME, FACILITIES};
